requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
google-generativeai==0.7.0
openai==1.30.1
python-dotenv==1.0.1
